* src/pathx.c: make numbers in path expressions 64 bit integers
Previously, they were whatever 'int' was, which might be too small, for
example to handle timestamps. With making them int64_t, it is now possible
to find files that have changed since a certain point in time with

  match /augeas/files//*[mtime][int(mtime) >= TIMESTAMP]

where TIMESTAMP is in seconds since the epoch

+++ src/pathx.c
@@ -184,7 +184,7 @@ struct value {
     enum type tag;
     union {
         struct nodeset  *nodeset;     /* T_NODESET */
-        int              number;      /* T_NUMBER  */
+        int64_t          number;      /* T_NUMBER  */
         char            *string;      /* T_STRING  */
         bool             boolval;     /* T_BOOLEAN */
         struct regexp   *regexp;      /* T_REGEXP  */
